Add icon font file
Used an icon font out of habit, although this time I generated a font
that's a subset of FontAwesome with only icons I'm using. I might
revisit this later as the render quality of SVGs is much higher, but at
least my implementation doesn't seem to trip up screen readers.